Module sui::client_commands

Structs

AddressesOutput

DynamicFieldOutput

GasCoinOutput

NewAddressOutput

ObjectOutput

ObjectsOutput

Opts

Global options for most transaction execution related commands

OptsWithGas

Global options with gas

SwitchResponse

Enums

SuiClientCommandResult

SuiClientCommands

Constants

GAS_SAFE_OVERHEAD

Only to be used within CLI

Functions

estimate_gas_budget

Call a dry run with the transaction data to estimate the gas budget. The estimated gas budget is computed as following:

estimate_gas_budget_from_gas_cost

execute_dry_run

Helper function to reduce code duplication for executing dry run

max_gas_budget

Queries the protocol config for the maximum gas allowed in a transaction.

request_tokens_from_faucet

Request tokens from the Faucet for the given address