Enum sui::client_commands::SuiClientCommands

pub enum SuiClientCommands {

    ActiveAddress,
    ActiveEnv,
    Addresses,
    Call {
        package: ObjectID,
        module: String,
        function: String,
        type_args: Vec<TypeTag>,
        args: Vec<SuiJsonValue>,
        gas: Option<ObjectID>,
        gas_budget: u64,
        serialize_unsigned_transaction: bool,
        serialize_signed_transaction: bool,
    },
    ChainIdentifier,
    DynamicFieldQuery {
        id: ObjectID,
        cursor: Option<ObjectID>,
        limit: usize,
    },
    Envs,
    ExecuteSignedTx {
        tx_bytes: String,
        signatures: Vec<String>,
    },
    Gas {
        address: Option<KeyIdentity>,
    },
    MergeCoin {
        primary_coin: ObjectID,
        coin_to_merge: ObjectID,
        gas: Option<ObjectID>,
        gas_budget: u64,
        serialize_unsigned_transaction: bool,
        serialize_signed_transaction: bool,
    },
    NewAddress {
        key_scheme: SignatureScheme,
        alias: Option<String>,
        word_length: Option<String>,
        derivation_path: Option<DerivationPath>,
    },
    NewEnv {
        alias: String,
        rpc: String,
        ws: Option<String>,
    },
    Object {
        id: ObjectID,
        bcs: bool,
    },
    Objects {
        address: Option<KeyIdentity>,
    },
    Pay {
        input_coins: Vec<ObjectID>,
        recipients: Vec<KeyIdentity>,
        amounts: Vec<u64>,
        gas: Option<ObjectID>,
        gas_budget: u64,
        serialize_unsigned_transaction: bool,
        serialize_signed_transaction: bool,
    },
    PayAllSui {
        input_coins: Vec<ObjectID>,
        recipient: KeyIdentity,
        gas_budget: u64,
        serialize_unsigned_transaction: bool,
        serialize_signed_transaction: bool,
    },
    PaySui {
        input_coins: Vec<ObjectID>,
        recipients: Vec<KeyIdentity>,
        amounts: Vec<u64>,
        gas_budget: u64,
        serialize_unsigned_transaction: bool,
        serialize_signed_transaction: bool,
    },
    Publish {
        package_path: PathBuf,
        build_config: MoveBuildConfig,
        gas: Option<ObjectID>,
        gas_budget: u64,
        skip_dependency_verification: bool,
        with_unpublished_dependencies: bool,
        serialize_unsigned_transaction: bool,
        serialize_signed_transaction: bool,
    },
    SplitCoin {
        coin_id: ObjectID,
        amounts: Option<Vec<u64>>,
        count: Option<u64>,
        gas: Option<ObjectID>,
        gas_budget: u64,
        serialize_unsigned_transaction: bool,
        serialize_signed_transaction: bool,
    },
    Switch {
        address: Option<KeyIdentity>,
        env: Option<String>,
    },
    TransactionBlock {
        digest: TransactionDigest,
    },
    Transfer {
        to: KeyIdentity,
        object_id: ObjectID,
        gas: Option<ObjectID>,
        gas_budget: u64,
        serialize_unsigned_transaction: bool,
        serialize_signed_transaction: bool,
    },
    TransferSui {
        to: KeyIdentity,
        sui_coin_object_id: ObjectID,
        gas_budget: u64,
        amount: Option<u64>,
        serialize_unsigned_transaction: bool,
        serialize_signed_transaction: bool,
    },
    Upgrade {
        package_path: PathBuf,
        upgrade_capability: ObjectID,
        build_config: MoveBuildConfig,
        gas: Option<ObjectID>,
        gas_budget: u64,
        skip_dependency_verification: bool,
        with_unpublished_dependencies: bool,
        serialize_unsigned_transaction: bool,
        serialize_signed_transaction: bool,
    },
    VerifyBytecodeMeter {
        package_path: PathBuf,
        build_config: MoveBuildConfig,
    },
    VerifySource {
        package_path: PathBuf,
        build_config: MoveBuildConfig,
        verify_deps: bool,
        skip_source: bool,
        address_override: Option<ObjectID>,
    },
    ReplayTransaction {
        tx_digest: String,
    },
    ReplayBatch {
        path: PathBuf,
        terminate_early: bool,
    },
    ReplayCheckpoints {
        start: u64,
        end: u64,
        terminate_early: bool,
    },

}

Variants

ActiveAddress

Default address used for commands when none specified

ActiveEnv

Default environment used for commands when none specified

Addresses

Obtain the Addresses managed by the client.

Call

Fields

package: ObjectID

Object ID of the package, which contains the module

module: String

The name of the module in the package

function: String

Function name in module

type_args: Vec<TypeTag>

Type arguments to the generic function being called. All must be specified, or the call will fail.

args: Vec<SuiJsonValue>

Simplified ordered args like in the function syntax ObjectIDs, Addresses must be hex strings

gas: Option<ObjectID>

ID of the gas object for gas payment, in 20 bytes Hex string If not provided, a gas object with at least gas_budget value will be selected

gas_budget: u64

Gas budget for this call

serialize_unsigned_transaction: bool

Instead of executing the transaction, serialize the bcs bytes of the unsigned transaction data (TransactionData) using base64 encoding, and print out the string.

serialize_signed_transaction: bool

Instead of executing the transaction, serialize the bcs bytes of the signed transaction data (SenderSignedData) using base64 encoding, and print out the string.

Call Move function

ChainIdentifier

Query the chain identifier from the rpc endpoint.

DynamicFieldQuery

Fields

id: ObjectID

The ID of the parent object

cursor: Option<ObjectID>

Optional paging cursor

limit: usize

Maximum item returned per page

Query a dynamic field by its address.

Envs

List all Sui environments

ExecuteSignedTx

Fields

tx_bytes: String

BCS serialized transaction data bytes without its type tag, as base-64 encoded string.

signatures: Vec<String>

A list of Base64 encoded signatures flag || signature || pubkey.

Execute a Signed Transaction. This is useful when the user prefers to sign elsewhere and use this command to execute.

Gas

Fields

address: Option<KeyIdentity>

Address (or its alias) owning the objects

Obtain all gas objects owned by the address. An address’ alias can be used instead of the address.

MergeCoin

Fields

primary_coin: ObjectID

The address of the coin to merge into.

coin_to_merge: ObjectID

The address of the coin to be merged.

gas: Option<ObjectID>

The address of the gas object for gas payment. If not provided, a gas object with at least gas_budget value will be selected.

gas_budget: u64

Gas budget for this call

serialize_unsigned_transaction: bool

Instead of executing the transaction, serialize the bcs bytes of the unsigned transaction data (TransactionData) using base64 encoding, and print out the string.

serialize_signed_transaction: bool

Instead of executing the transaction, serialize the bcs bytes of the signed transaction data (SenderSignedData) using base64 encoding, and print out the string.

Merge two coin objects into one coin

NewAddress

Fields

key_scheme: SignatureScheme

alias: Option<String>

The alias must start with a letter and can contain only letters, digits, hyphens (-), or underscores (_).

word_length: Option<String>

derivation_path: Option<DerivationPath>

Generate new address and keypair with keypair scheme flag {ed25519 | secp256k1 | secp256r1} with optional derivation path, default to m/44’/784’/0’/0’/0’ for ed25519 or m/54’/784’/0’/0/0 for secp256k1 or m/74’/784’/0’/0/0 for secp256r1. Word length can be { word12 | word15 | word18 | word21 | word24} default to word12 if not specified.

NewEnv

Fields

alias: String

rpc: String

ws: Option<String>

Add new Sui environment.

Object

Fields

id: ObjectID

Object ID of the object to fetch

bcs: bool

Return the bcs serialized version of the object

Get object info

Objects

Fields

address: Option<KeyIdentity>

Address owning the object. If no address is provided, it will show all objects owned by sui client active-address.

Obtain all objects owned by the address. It also accepts an address by its alias.

Pay

Fields

input_coins: Vec<ObjectID>

The input coins to be used for pay recipients, following the specified amounts.

recipients: Vec<KeyIdentity>

The recipient addresses, must be of same length as amounts Aliases of addresses are also accepted as input.

amounts: Vec<u64>

The amounts to be paid, following the order of recipients.

gas: Option<ObjectID>

ID of the gas object for gas payment, in 20 bytes Hex string If not provided, a gas object with at least gas_budget value will be selected

gas_budget: u64

Gas budget for this transaction

serialize_unsigned_transaction: bool

Instead of executing the transaction, serialize the bcs bytes of the unsigned transaction data (TransactionData) using base64 encoding, and print out the string.

serialize_signed_transaction: bool

Instead of executing the transaction, serialize the bcs bytes of the signed transaction data (SenderSignedData) using base64 encoding, and print out the string.

Pay coins to recipients following specified amounts, with input coins. Length of recipients must be the same as that of amounts.

PayAllSui

Fields

input_coins: Vec<ObjectID>

The input coins to be used for pay recipients, including the gas coin.

recipient: KeyIdentity

The recipient address (or its alias if it’s an address in the keystore).

gas_budget: u64

Gas budget for this transaction

serialize_unsigned_transaction: bool

Instead of executing the transaction, serialize the bcs bytes of the unsigned transaction data (TransactionData) using base64 encoding, and print out the string.

serialize_signed_transaction: bool

Instead of executing the transaction, serialize the bcs bytes of the signed transaction data (SenderSignedData) using base64 encoding, and print out the string.

Pay all residual SUI coins to the recipient with input coins, after deducting the gas cost. The input coins also include the coin for gas payment, so no extra gas coin is required.

PaySui

Fields

input_coins: Vec<ObjectID>

The input coins to be used for pay recipients, including the gas coin.

recipients: Vec<KeyIdentity>

The recipient addresses, must be of same length as amounts. Aliases of addresses are also accepted as input.

amounts: Vec<u64>

The amounts to be paid, following the order of recipients.

gas_budget: u64

Gas budget for this transaction

serialize_unsigned_transaction: bool

Instead of executing the transaction, serialize the bcs bytes of the unsigned transaction data (TransactionData) using base64 encoding, and print out the string.

serialize_signed_transaction: bool

Instead of executing the transaction, serialize the bcs bytes of the signed transaction data (SenderSignedData) using base64 encoding, and print out the string.

Pay SUI coins to recipients following following specified amounts, with input coins. Length of recipients must be the same as that of amounts. The input coins also include the coin for gas payment, so no extra gas coin is required.

Publish

Fields

package_path: PathBuf

Path to directory containing a Move package

build_config: MoveBuildConfig

Package build options

gas: Option<ObjectID>

ID of the gas object for gas payment, in 20 bytes Hex string If not provided, a gas object with at least gas_budget value will be selected

gas_budget: u64

Gas budget for running module initializers

skip_dependency_verification: bool

Publish the package without checking whether compiling dependencies from source results in bytecode matching the dependencies found on-chain.

with_unpublished_dependencies: bool

Also publish transitive dependencies that have not already been published.

serialize_unsigned_transaction: bool

Instead of executing the transaction, serialize the bcs bytes of the unsigned transaction data (TransactionData) using base64 encoding, and print out the string.

serialize_signed_transaction: bool

Instead of executing the transaction, serialize the bcs bytes of the signed transaction data (SenderSignedData) using base64 encoding, and print out the string.

Publish Move modules

SplitCoin

Fields

coin_id: ObjectID

Coin to Split, in 20 bytes Hex string

amounts: Option<Vec<u64>>

Specific amounts to split out from the coin

count: Option<u64>

Count of equal-size coins to split into

gas: Option<ObjectID>

ID of the gas object for gas payment, in 20 bytes Hex string If not provided, a gas object with at least gas_budget value will be selected

gas_budget: u64

Gas budget for this call

serialize_unsigned_transaction: bool

Instead of executing the transaction, serialize the bcs bytes of the unsigned transaction data (TransactionData) using base64 encoding, and print out the string.

serialize_signed_transaction: bool

Instead of executing the transaction, serialize the bcs bytes of the signed transaction data (SenderSignedData) using base64 encoding, and print out the string.

Split a coin object into multiple coins.

Switch

Fields

address: Option<KeyIdentity>

An address to be used as the active address for subsequent commands. It accepts also the alias of the address.

env: Option<String>

The RPC server URL (e.g., local rpc server, devnet rpc server, etc) to be used for subsequent commands.

Switch active address and network(e.g., devnet, local rpc server).

TransactionBlock

Fields

digest: TransactionDigest

Digest of the transaction block

Get the effects of executing the given transaction block

Transfer

Fields

to: KeyIdentity

Recipient address (or its alias if it’s an address in the keystore)

object_id: ObjectID

Object to transfer, in 20 bytes Hex string

gas: Option<ObjectID>

ID of the gas object for gas payment, in 20 bytes Hex string If not provided, a gas object with at least gas_budget value will be selected

gas_budget: u64

Gas budget for this transfer

serialize_unsigned_transaction: bool

Instead of executing the transaction, serialize the bcs bytes of the unsigned transaction data (TransactionData) using base64 encoding, and print out the string.

serialize_signed_transaction: bool

Instead of executing the transaction, serialize the bcs bytes of the signed transaction data (SenderSignedData) using base64 encoding, and print out the string.

Transfer object

TransferSui

Fields

to: KeyIdentity

Recipient address (or its alias if it’s an address in the keystore)

sui_coin_object_id: ObjectID

Sui coin object to transfer, ID in 20 bytes Hex string. This is also the gas object.

gas_budget: u64

Gas budget for this transfer

amount: Option<u64>

The amount to transfer, if not specified, the entire coin object will be transferred.

serialize_unsigned_transaction: bool

Instead of executing the transaction, serialize the bcs bytes of the unsigned transaction data (TransactionData) using base64 encoding, and print out the string.

serialize_signed_transaction: bool

Instead of executing the transaction, serialize the bcs bytes of the signed transaction data (SenderSignedData) using base64 encoding, and print out the string.

Transfer SUI, and pay gas with the same SUI coin object. If amount is specified, only the amount is transferred; otherwise the entire object is transferred.

Upgrade

Fields

package_path: PathBuf

Path to directory containing a Move package

upgrade_capability: ObjectID

ID of the upgrade capability for the package being upgraded.

build_config: MoveBuildConfig

Package build options

gas: Option<ObjectID>

ID of the gas object for gas payment, in 20 bytes Hex string If not provided, a gas object with at least gas_budget value will be selected

gas_budget: u64

Gas budget for running module initializers

skip_dependency_verification: bool

Publish the package without checking whether compiling dependencies from source results in bytecode matching the dependencies found on-chain.

with_unpublished_dependencies: bool

Also publish transitive dependencies that have not already been published.

serialize_unsigned_transaction: bool

Instead of executing the transaction, serialize the bcs bytes of the unsigned transaction data (TransactionData) using base64 encoding, and print out the string.

serialize_signed_transaction: bool

Instead of executing the transaction, serialize the bcs bytes of the signed transaction data (SenderSignedData) using base64 encoding, and print out the string.

Upgrade Move modules

VerifyBytecodeMeter

Fields

package_path: PathBuf

Path to directory containing a Move package

build_config: MoveBuildConfig

Package build options

Run the bytecode verifier on the package

VerifySource

Fields

package_path: PathBuf

Path to directory containing a Move package

build_config: MoveBuildConfig

Package build options

verify_deps: bool

Verify on-chain dependencies.

skip_source: bool

Don’t verify source (only valid if –verify-deps is enabled).

address_override: Option<ObjectID>

If specified, override the addresses for the package’s own modules with this address. Only works for unpublished modules (whose addresses are currently 0x0).

Verify local Move packages against on-chain packages, and optionally their dependencies.

ReplayTransaction

Fields

tx_digest: String

The digest of the transaction to replay

Replay a given transaction to view transaction effects. Set environment variable MOVE_VM_STEP=1 to debug.

ReplayBatch

Fields

path: PathBuf

The path to the file of transaction digests to replay, with one digest per line

terminate_early: bool

If an error is encountered during a transaction, this specifies whether to terminate or continue

Replay transactions listed in a file.

ReplayCheckpoints

Fields

start: u64

The starting checkpoint sequence number of the range of checkpoints to replay

end: u64

The ending checkpoint sequence number of the range of checkpoints to replay

terminate_early: bool

If an error is encountered during a transaction, this specifies whether to terminate or continue

Replay all transactions in a range of checkpoints.

Implementations

impl SuiClientCommands

pub async fn execute( self, context: &mut WalletContext ) -> Result<SuiClientCommandResult, Error>

pub fn switch_env(config: &mut SuiClientConfig, env: &str) -> Result<(), Error>