Module sui::client_ptb::ast

Structs

ModuleAccess

A parsed module access consisting of the address, module name, and function name.

Program

A PTB Program consisting of a list of commands and a flag indicating if the preview warn-shadows command was present.

ProgramMetadata

The ProgramMetadata struct holds metadata about a PTB program, such as whether the preview flag was set, json output was set, etc.

Enums

Argument

An enum representing the parsed arguments of a PTB command.

ParsedPTBCommand

A parsed PTB command consisting of the command and the parsed arguments to the command.

Constants

ADDRESS

ASSIGN

BOOL

COMMANDS

DRY_RUN

GAS

GAS_BUDGET

GAS_COIN

JSON

KEYWORDS

MAKE_MOVE_VEC

MERGE_COINS

MOVE_CALL

NONE

PREVIEW

PUBLISH

SERIALIZE_SIGNED

SERIALIZE_UNSIGNED

SOME

SPLIT_COINS

SUMMARY

TRANSFER_OBJECTS

U8

U16

U32

U64

U128

U256

UPGRADE

VECTOR

WARN_SHADOWS

Functions

all_keywords

is_keyword

Type Aliases

ParsedProgram