sui::client_ptb

Module error

Macros

sp

Structs

PTBError

An error with a message, a location in the source code, and an optional help message.

Span

Represents the location of a range of text in the PTB source.

Spanned

A value that has an associated location in source code.

Functions

build_error_reports

Type Aliases

PTBResult