sui::client_ptb::ptb

Function to_source_string

pub fn to_source_string(strings: Vec<String>) -> String

Convert a vector of shell tokens into a single string, with each shell token separated by a space with each command starting on a new line. NB: we add a space to the end of the source string to ensure that for unexpected EOF errors we have a location to point to.