Enum sui::keytool::KeyToolCommand

pub enum KeyToolCommand {

    Generate {
        key_scheme: SignatureScheme,
        word_length: Option<String>,
        derivation_path: Option<DerivationPath>,
    },
    Show {
        file: PathBuf,
    },
    Unpack {
        keypair: SuiKeyPair,
    },
    List,
    Sign {
        address: SuiAddress,
        data: String,
        intent: Option<Intent>,
    },
    SignKMS {
        data: String,
        keyid: String,
        intent: Option<Intent>,
        base64pk: String,
    },
    Import {
        mnemonic_phrase: String,
        key_scheme: SignatureScheme,
        derivation_path: Option<DerivationPath>,
    },
    Convert {
        value: String,
    },
    LoadKeypair {
        file: PathBuf,
    },
    Base64PubKeyToAddress {
        base64_key: String,
    },
    MultiSigAddress {
        threshold: ThresholdUnit,
        pks: Vec<PublicKey>,
        weights: Vec<WeightUnit>,
    },
    MultiSigCombinePartialSig {
        sigs: Vec<Signature>,
        pks: Vec<PublicKey>,
        weights: Vec<WeightUnit>,
        threshold: ThresholdUnit,
    },
    DecodeMultiSig {
        multisig: MultiSig,
    },
    DecodeTxBytes {
        tx_bytes: String,
    },
    Base64ToHex {
        base64: String,
    },
    HexToBase64 {
        hex: String,
    },
    HexToBytes {
        hex: String,
    },
    BytesToHex {
        bytes: Vec<u8>,
    },
    Base64ToBytes {
        base64: String,
    },
    BytesToBase64 {
        bytes: Vec<u8>,
    },

}

Variants

Generate

Fields

key_scheme: SignatureScheme

word_length: Option<String>

derivation_path: Option<DerivationPath>

Generate a new keypair with key scheme flag {ed25519 | secp256k1 | secp256r1} with optional derivation path, default to m/44’/784’/0’/0’/0’ for ed25519 or m/54’/784’/0’/0/0 for secp256k1 or m/74’/784’/0’/0/0 for secp256r1. Word length can be { word12 | word15 | word18 | word21 | word24} default to word12 if not specified.

The keypair file is output to the current directory. The content of the file is a Base64 encoded string of 33-byte flag || privkey. Note: To generate and add keypair to sui.keystore, use sui client new-address), see more at [enum SuiClientCommands].

Show

Fields

file: PathBuf

This reads the content at the provided file path. The accepted format can be [enum SuiKeyPair] (Base64 encoded of 33-byte flag || privkey) or type AuthorityKeyPair (Base64 encoded privkey). It prints its Base64 encoded public key and the key scheme flag.

Unpack

Fields

keypair: SuiKeyPair

This takes [enum SuiKeyPair] of Base64 encoded of 33-byte flag || privkey). It outputs the keypair into a file at the current directory, and prints out its Sui address, Base64 encoded public key, and the key scheme flag.

List

List all keys by its Sui address, Base64 encoded public key, key scheme name in sui.keystore.

Sign

Fields

address: SuiAddress

data: String

intent: Option<Intent>

Create signature using the private key for for the given address in sui keystore. Any signature commits to a [struct IntentMessage] consisting of the Base64 encoded of the BCS serialized transaction bytes itself (the result of transaction builder API and its intent. If intent is absent, default will be used. See [struct IntentMessage] and [struct Intent] for more details.

SignKMS

Fields

data: String

keyid: String

intent: Option<Intent>

base64pk: String

Creates a signature by leveraging AWS KMS. Pass in a key-id to leverage Amazon KMS to sign a message and the base64 pubkey. Generate PubKey from pem using MystenLabs/base64pemkey Any signature commits to a [struct IntentMessage] consisting of the Base64 encoded of the BCS serialized transaction bytes itself (the result of transaction builder API and its intent. If intent is absent, default will be used. See [struct IntentMessage] and [struct Intent] for more details.

Import

Fields

mnemonic_phrase: String

key_scheme: SignatureScheme

derivation_path: Option<DerivationPath>

Add a new key to sui.keystore based on the input mnemonic phrase, the key scheme flag {ed25519 | secp256k1 | secp256r1} and an optional derivation path, default to m/44’/784’/0’/0’/0’ for ed25519 or m/54’/784’/0’/0/0 for secp256k1 or m/74’/784’/0’/0/0 for secp256r1. Supports mnemonic phrase of word length 12, 15, 18`, 21, 24.

Convert

Fields

value: String

Convert private key from wallet format (hex of 32 byte private key) to sui.keystore format (base64 of 33 byte flag || private key) or vice versa.

LoadKeypair

Fields

file: PathBuf

This reads the content at the provided file path. The accepted format can be [enum SuiKeyPair] (Base64 encoded of 33-byte flag || privkey) or type AuthorityKeyPair (Base64 encoded privkey). This prints out the account keypair as Base64 encoded flag || privkey, the network keypair, worker keypair, protocol keypair as Base64 encoded privkey.

Base64PubKeyToAddress

Fields

base64_key: String

MultiSigAddress

Fields

threshold: ThresholdUnit

pks: Vec<PublicKey>

weights: Vec<WeightUnit>

To MultiSig Sui Address. Pass in a list of all public keys flag || pk in Base64. See keytool list for example public keys.

MultiSigCombinePartialSig

Fields

sigs: Vec<Signature>

pks: Vec<PublicKey>

weights: Vec<WeightUnit>

threshold: ThresholdUnit

Provides a list of signatures (flag || sig || pk encoded in Base64), threshold, a list of public keys. Returns a valid MultiSig and its sender address. The result can be used as signature field for sui client execute-signed-tx. The number of sigs must be greater than the threshold. The number of sigs must be smaller than the number of pks.

DecodeMultiSig

Fields

multisig: MultiSig

Given a Base64 encoded MultiSig signature, decode its components.

DecodeTxBytes

Fields

tx_bytes: String

Given a Base64 encoded transaction bytes, decode its components.

Base64ToHex

Fields

base64: String

Converts a Base64 encoded string to its hexadecimal representation.

HexToBase64

Fields

hex: String

Converts a hexadecimal string to its Base64 encoded representation.

HexToBytes

Fields

hex: String

Converts a hexadecimal string to its corresponding bytes.

BytesToHex

Fields

bytes: Vec<u8>

Converts an array of bytes to its hexadecimal string representation.

Base64ToBytes

Fields

base64: String

Decodes a Base64 encoded string to its corresponding bytes.

BytesToBase64

Fields

bytes: Vec<u8>

Encodes an array of bytes to its Base64 string representation.

Implementations

impl KeyToolCommand

pub async fn execute(self, keystore: &mut Keystore) -> Result<(), Error>