Trait sui::rpc_gateway::RpcGatewayServer

pub trait RpcGatewayServer: Sized + Send + Sync + 'static {

    fn get_object_typed_info<'life0, 'async_trait>(
        &'life0 self, 
        object_id: ObjectID
    ) -> Pin<Box<dyn Future<Output = RpcResult<GetObjectInfoResponse>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn transfer_coin<'life0, 'async_trait>(
        &'life0 self, 
        signer: SuiAddress, 
        object_id: ObjectID, 
        gas_payment: ObjectID, 
        gas_budget: u64, 
        recipient: SuiAddress
    ) -> Pin<Box<dyn Future<Output = RpcResult<TransactionBytes>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn move_call<'life0, 'async_trait>(
        &'life0 self, 
        signer: SuiAddress, 
        package_object_id: ObjectID, 
        module: Identifier, 
        function: Identifier, 
        type_arguments: Vec<SuiTypeTag>, 
        arguments: Vec<SuiJsonValue>, 
        gas_object_id: ObjectID, 
        gas_budget: u64
    ) -> Pin<Box<dyn Future<Output = RpcResult<TransactionBytes>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn publish<'life0, 'async_trait>(
        &'life0 self, 
        sender: SuiAddress, 
        compiled_modules: Vec<Base64>, 
        gas_object_id: ObjectID, 
        gas_budget: u64
    ) -> Pin<Box<dyn Future<Output = RpcResult<TransactionBytes>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn split_coin<'life0, 'async_trait>(
        &'life0 self, 
        signer: SuiAddress, 
        coin_object_id: ObjectID, 
        split_amounts: Vec<u64>, 
        gas_payment: ObjectID, 
        gas_budget: u64
    ) -> Pin<Box<dyn Future<Output = RpcResult<TransactionBytes>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn merge_coin<'life0, 'async_trait>(
        &'life0 self, 
        signer: SuiAddress, 
        primary_coin: ObjectID, 
        coin_to_merge: ObjectID, 
        gas_payment: ObjectID, 
        gas_budget: u64
    ) -> Pin<Box<dyn Future<Output = RpcResult<TransactionBytes>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn execute_transaction<'life0, 'async_trait>(
        &'life0 self, 
        signed_transaction: SignedTransaction
    ) -> Pin<Box<dyn Future<Output = RpcResult<TransactionResponse>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn sync_account_state<'life0, 'async_trait>(
        &'life0 self, 
        address: SuiAddress
    ) -> Pin<Box<dyn Future<Output = RpcResult<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_owned_objects<'life0, 'async_trait>(
        &'life0 self, 
        owner: SuiAddress
    ) -> Pin<Box<dyn Future<Output = RpcResult<ObjectResponse>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_total_transaction_number<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = RpcResult<u64>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_transactions_in_range<'life0, 'async_trait>(
        &'life0 self, 
        start: GatewayTxSeqNumber, 
        end: GatewayTxSeqNumber
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<(GatewayTxSeqNumber, TransactionDigest)>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_recent_transactions<'life0, 'async_trait>(
        &'life0 self, 
        count: u64
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<(GatewayTxSeqNumber, TransactionDigest)>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_transaction<'life0, 'async_trait>(
        &'life0 self, 
        digest: TransactionDigest
    ) -> Pin<Box<dyn Future<Output = RpcResult<TransactionEffectsResponse>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_object_info<'life0, 'async_trait>(
        &'life0 self, 
        object_id: ObjectID
    ) -> Pin<Box<dyn Future<Output = RpcResult<ObjectRead>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn into_rpc(self) -> RpcModule<Self> { ... }

}

Server trait implementation for the RpcGateway RPC API.

Required Methods

fn get_object_typed_info<'life0, 'async_trait>(
&'life0 self,
object_id: ObjectID
) -> Pin<Box<dyn Future<Output = RpcResult<GetObjectInfoResponse>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Return the object information for a specified object

fn transfer_coin<'life0, 'async_trait>(
&'life0 self,
signer: SuiAddress,
object_id: ObjectID,
gas_payment: ObjectID,
gas_budget: u64,
recipient: SuiAddress
) -> Pin<Box<dyn Future<Output = RpcResult<TransactionBytes>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Create a transaction to transfer a Sui coin from one address to another.

fn move_call<'life0, 'async_trait>(
&'life0 self,
signer: SuiAddress,
package_object_id: ObjectID,
module: Identifier,
function: Identifier,
type_arguments: Vec<SuiTypeTag>,
arguments: Vec<SuiJsonValue>,
gas_object_id: ObjectID,
gas_budget: u64
) -> Pin<Box<dyn Future<Output = RpcResult<TransactionBytes>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Execute a Move call transaction by calling the specified function in the module of a given package.

fn publish<'life0, 'async_trait>(
&'life0 self,
sender: SuiAddress,
compiled_modules: Vec<Base64>,
gas_object_id: ObjectID,
gas_budget: u64
) -> Pin<Box<dyn Future<Output = RpcResult<TransactionBytes>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Publish Move module.

fn split_coin<'life0, 'async_trait>(
&'life0 self,
signer: SuiAddress,
coin_object_id: ObjectID,
split_amounts: Vec<u64>,
gas_payment: ObjectID,
gas_budget: u64
) -> Pin<Box<dyn Future<Output = RpcResult<TransactionBytes>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

fn merge_coin<'life0, 'async_trait>(
&'life0 self,
signer: SuiAddress,
primary_coin: ObjectID,
coin_to_merge: ObjectID,
gas_payment: ObjectID,
gas_budget: u64
) -> Pin<Box<dyn Future<Output = RpcResult<TransactionBytes>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

fn execute_transaction<'life0, 'async_trait>(
&'life0 self,
signed_transaction: SignedTransaction
) -> Pin<Box<dyn Future<Output = RpcResult<TransactionResponse>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Execute the transaction using the transaction data, signature and public key.

fn sync_account_state<'life0, 'async_trait>(
&'life0 self,
address: SuiAddress
) -> Pin<Box<dyn Future<Output = RpcResult<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Synchronize client state with validators.

fn get_owned_objects<'life0, 'async_trait>(
&'life0 self,
owner: SuiAddress
) -> Pin<Box<dyn Future<Output = RpcResult<ObjectResponse>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Return the list of objects owned by an address.

fn get_total_transaction_number<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = RpcResult<u64>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

fn get_transactions_in_range<'life0, 'async_trait>(
&'life0 self,
start: GatewayTxSeqNumber,
end: GatewayTxSeqNumber
) -> Pin<Box<dyn Future<Output = RpcResult<Vec<(GatewayTxSeqNumber, TransactionDigest)>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

fn get_recent_transactions<'life0, 'async_trait>(
&'life0 self,
count: u64
) -> Pin<Box<dyn Future<Output = RpcResult<Vec<(GatewayTxSeqNumber, TransactionDigest)>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

fn get_transaction<'life0, 'async_trait>(
&'life0 self,
digest: TransactionDigest
) -> Pin<Box<dyn Future<Output = RpcResult<TransactionEffectsResponse>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

fn get_object_info<'life0, 'async_trait>(
&'life0 self,
object_id: ObjectID
) -> Pin<Box<dyn Future<Output = RpcResult<ObjectRead>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Low level API to get object info. Client Applications should prefer to use get_object_typed_info instead.

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.