Enum sui::wallet_commands::WalletCommands

pub enum WalletCommands {

    Switch {
        address: Option<SuiAddress>,
        gateway: Option<String>,
    },
    ActiveAddress {},
    Object {
        id: ObjectID,
    },
    Publish {
        path: String,
        gas: Option<ObjectID>,
        gas_budget: u64,
    },
    Call {
        package: ObjectID,
        module: Identifier,
        function: Identifier,
        type_args: Vec<TypeTag>,
        args: Vec<SuiJsonValue>,
        gas: Option<ObjectID>,
        gas_budget: u64,
    },
    Transfer {
        to: SuiAddress,
        coin_object_id: ObjectID,
        gas: Option<ObjectID>,
        gas_budget: u64,
    },
    SyncClientState {
        address: Option<SuiAddress>,
    },
    Addresses,
    NewAddress,
    Objects {
        address: Option<SuiAddress>,
    },
    Gas {
        address: Option<SuiAddress>,
    },
    SplitCoin {
        coin_id: ObjectID,
        amounts: Vec<u64>,
        gas: Option<ObjectID>,
        gas_budget: u64,
    },
    MergeCoin {
        primary_coin: ObjectID,
        coin_to_merge: ObjectID,
        gas: Option<ObjectID>,
        gas_budget: u64,
    },
    CreateExampleNFT {
        name: Option<String>,
        description: Option<String>,
        url: Option<String>,
        gas: Option<ObjectID>,
        gas_budget: Option<u64>,
    },

}

Variants

Switch

Fields

address: Option<SuiAddress>

An Sui address to be used as the active address for subsequent commands.

gateway: Option<String>

The gateway URL (e.g., local rpc server, devnet rpc server, etc) to be used for subsequent commands.

Switch active address and network(e.g., devnet, local rpc server)

ActiveAddress

Fields

Default address used for commands when none specified

Object

Fields

id: ObjectID

Object ID of the object to fetch

Get obj info

Publish

Fields

path: String

Path to directory containing a Move package

gas: Option<ObjectID>

ID of the gas object for gas payment, in 20 bytes Hex string If not provided, a gas object with at least gas_budget value will be selected

gas_budget: u64

Gas budget for running module initializers

Publish Move modules

Call

Fields

package: ObjectID

Object ID of the package, which contains the module

module: Identifier

The name of the module in the package

function: Identifier

Function name in module

type_args: Vec<TypeTag>

Function name in module

args: Vec<SuiJsonValue>

Simplified ordered args like in the function syntax ObjectIDs, Addresses must be hex strings

gas: Option<ObjectID>

ID of the gas object for gas payment, in 20 bytes Hex string If not provided, a gas object with at least gas_budget value will be selected

gas_budget: u64

Gas budget for this call

Call Move function

Transfer

Fields

to: SuiAddress

Recipient address

coin_object_id: ObjectID

Coin to transfer, in 20 bytes Hex string

gas: Option<ObjectID>

ID of the gas object for gas payment, in 20 bytes Hex string If not provided, a gas object with at least gas_budget value will be selected

gas_budget: u64

Gas budget for this transfer

Transfer coin object

SyncClientState

Fields

address: Option<SuiAddress>

Synchronize client state with authorities.

Addresses

Obtain the Addresses managed by the wallet.

NewAddress

Generate new address and keypair.

Objects

Fields

address: Option<SuiAddress>

Address owning the objects

Obtain all objects owned by the address.

Gas

Fields

address: Option<SuiAddress>

Address owning the objects

Obtain all gas objects owned by the address.

SplitCoin

Fields

coin_id: ObjectID

Coin to Split, in 20 bytes Hex string

amounts: Vec<u64>

Amount to split out from the coin

gas: Option<ObjectID>

ID of the gas object for gas payment, in 20 bytes Hex string If not provided, a gas object with at least gas_budget value will be selected

gas_budget: u64

Gas budget for this call

Split a coin object into multiple coins.

MergeCoin

Fields

primary_coin: ObjectID

Coin to merge into, in 20 bytes Hex string

coin_to_merge: ObjectID

Coin to be merged, in 20 bytes Hex string

gas: Option<ObjectID>

ID of the gas object for gas payment, in 20 bytes Hex string If not provided, a gas object with at least gas_budget value will be selected

gas_budget: u64

Gas budget for this call

Merge two coin objects into one coin

CreateExampleNFT

Fields

name: Option<String>

Name of the NFT

description: Option<String>

Description of the NFT

url: Option<String>

Display url(e.g., an image url) of the NFT

gas: Option<ObjectID>

ID of the gas object for gas payment, in 20 bytes Hex string If not provided, a gas object with at least gas_budget value will be selected

gas_budget: Option<u64>

Gas budget for this transfer

Create an example NFT

Implementations

impl WalletCommands

pub async fn execute(
&mut self,
context: &mut WalletContext
) -> Result<WalletCommandResult, Error>