sui
Module client_commands

Structs

AddressesOutput

DynamicFieldOutput

GasCoinOutput

GasDataArgs

Arguments related to setting gas data, apart from payment coins.

NewAddressOutput

ObjectOutput

ObjectsOutput

PaymentArgs

Arguments related to providing coins for gas payment

RemoveAddressOutput

SwitchResponse

TxProcessingArgs

Arguments related to what to do to a transaction after it has been built.

Enums

SuiClientCommandResult

SuiClientCommands

Constants

GAS_SAFE_OVERHEAD

Only to be used within CLI

Functions

estimate_gas_budget

Call a dry run with the transaction data to estimate the gas budget. The estimated gas budget is computed as following:

estimate_gas_budget_from_gas_cost

execute_dry_run

Helper function to reduce code duplication for executing dry run

max_gas_budget

Queries the protocol config for the maximum gas allowed in a transaction.

request_tokens_from_faucet

Request tokens from the Faucet for the given address