Enum sui::client_commands::SuiClientCommands

pub enum SuiClientCommands {

    Switch {
        address: Option<SuiAddress>,
        env: Option<String>,
    },
    NewEnv {
        alias: String,
        rpc: String,
        ws: Option<String>,
    },
    Envs,
    ActiveAddress,
    ActiveEnv,
    Object {
        id: ObjectID,
        bcs: bool,
    },
    Publish {
        package_path: PathBuf,
        build_config: MoveBuildConfig,
        gas: Option<ObjectID>,
        gas_budget: u64,
        verify_dependencies: bool,
    },
    VerifySource {
        package_path: PathBuf,
        build_config: MoveBuildConfig,
        verify_deps: bool,
        skip_source: bool,
        address_override: Option<ObjectID>,
    },
    Call {
        package: ObjectID,
        module: String,
        function: String,
        type_args: Vec<TypeTag>,
        args: Vec<SuiJsonValue>,
        gas: Option<ObjectID>,
        gas_budget: u64,
    },
    Transfer {
        to: SuiAddress,
        object_id: ObjectID,
        gas: Option<ObjectID>,
        gas_budget: u64,
    },
    TransferSui {
        to: SuiAddress,
        sui_coin_object_id: ObjectID,
        gas_budget: u64,
        amount: Option<u64>,
    },
    Pay {
        input_coins: Vec<ObjectID>,
        recipients: Vec<SuiAddress>,
        amounts: Vec<u64>,
        gas: Option<ObjectID>,
        gas_budget: u64,
    },
    PaySui {
        input_coins: Vec<ObjectID>,
        recipients: Vec<SuiAddress>,
        amounts: Vec<u64>,
        gas_budget: u64,
    },
    PayAllSui {
        input_coins: Vec<ObjectID>,
        recipient: SuiAddress,
        gas_budget: u64,
    },
    Addresses,
    NewAddress {
        key_scheme: SignatureScheme,
        derivation_path: Option<DerivationPath>,
    },
    Objects {
        address: Option<SuiAddress>,
    },
    Gas {
        address: Option<SuiAddress>,
    },
    SplitCoin {
        coin_id: ObjectID,
        amounts: Option<Vec<u64>>,
        count: Option<u64>,
        gas: Option<ObjectID>,
        gas_budget: u64,
    },
    MergeCoin {
        primary_coin: ObjectID,
        coin_to_merge: ObjectID,
        gas: Option<ObjectID>,
        gas_budget: u64,
    },
    CreateExampleNFT {
        name: Option<String>,
        description: Option<String>,
        url: Option<String>,
        gas: Option<ObjectID>,
        gas_budget: Option<u64>,
    },
    SerializeTransferSui {
        to: SuiAddress,
        sui_coin_object_id: ObjectID,
        gas_budget: u64,
        amount: Option<u64>,
    },
    ExecuteSignedTx {
        tx_bytes: String,
        signature: String,
    },

}

Variants

Switch

Fields

address: Option<SuiAddress>

An Sui address to be used as the active address for subsequent commands.

env: Option<String>

The RPC server URL (e.g., local rpc server, devnet rpc server, etc) to be used for subsequent commands.

Switch active address and network(e.g., devnet, local rpc server)

NewEnv

Fields

alias: String

rpc: String

ws: Option<String>

Add new Sui environment.

Envs

List all Sui environments

ActiveAddress

Default address used for commands when none specified

ActiveEnv

Default environment used for commands when none specified

Object

Fields

id: ObjectID

Object ID of the object to fetch

bcs: bool

Return the bcs serialized version of the object

Get object info

Publish

Fields

package_path: PathBuf

Path to directory containing a Move package

build_config: MoveBuildConfig

Package build options

gas: Option<ObjectID>

ID of the gas object for gas payment, in 20 bytes Hex string If not provided, a gas object with at least gas_budget value will be selected

gas_budget: u64

Gas budget for running module initializers

verify_dependencies: bool

Confirms that compiling dependencies from source results in bytecode matching the dependency found on-chain.

Publish Move modules

VerifySource

Fields

package_path: PathBuf

Path to directory containing a Move package

build_config: MoveBuildConfig

Package build options

verify_deps: bool

Verify on-chain dependencies.

skip_source: bool

Don’t verify source (only valid if –verify-deps is enabled).

address_override: Option<ObjectID>

If specified, override the addresses for the package’s own modules with this address. Only works for unpublished modules (whose addresses are currently 0x0).

Verify local Move packages against on-chain packages, and optionally their dependencies.

Call

Fields

package: ObjectID

Object ID of the package, which contains the module

module: String

The name of the module in the package

function: String

Function name in module

type_args: Vec<TypeTag>

Function name in module

args: Vec<SuiJsonValue>

Simplified ordered args like in the function syntax ObjectIDs, Addresses must be hex strings

gas: Option<ObjectID>

ID of the gas object for gas payment, in 20 bytes Hex string If not provided, a gas object with at least gas_budget value will be selected

gas_budget: u64

Gas budget for this call

Call Move function

Transfer

Fields

to: SuiAddress

Recipient address

object_id: ObjectID

Object to transfer, in 20 bytes Hex string

gas: Option<ObjectID>

ID of the gas object for gas payment, in 20 bytes Hex string If not provided, a gas object with at least gas_budget value will be selected

gas_budget: u64

Gas budget for this transfer

Transfer object

TransferSui

Fields

to: SuiAddress

Recipient address

sui_coin_object_id: ObjectID

Sui coin object to transfer, ID in 20 bytes Hex string. This is also the gas object.

gas_budget: u64

Gas budget for this transfer

amount: Option<u64>

The amount to transfer, if not specified, the entire coin object will be transferred.

Transfer SUI, and pay gas with the same SUI coin object. If amount is specified, only the amount is transferred; otherwise the entire object is transferred.

Pay

Fields

input_coins: Vec<ObjectID>

The input coins to be used for pay recipients, following the specified amounts.

recipients: Vec<SuiAddress>

The recipient addresses, must be of same length as amounts

amounts: Vec<u64>

The amounts to be paid, following the order of recipients.

gas: Option<ObjectID>

ID of the gas object for gas payment, in 20 bytes Hex string If not provided, a gas object with at least gas_budget value will be selected

gas_budget: u64

Gas budget for this transaction

Pay coins to recipients following specified amounts, with input coins. Length of recipients must be the same as that of amounts.

PaySui

Fields

input_coins: Vec<ObjectID>

The input coins to be used for pay recipients, including the gas coin.

recipients: Vec<SuiAddress>

The recipient addresses, must be of same length as amounts.

amounts: Vec<u64>

The amounts to be paid, following the order of recipients.

gas_budget: u64

Gas budget for this transaction

Pay SUI coins to recipients following following specified amounts, with input coins. Length of recipients must be the same as that of amounts. The input coins also include the coin for gas payment, so no extra gas coin is required.

PayAllSui

Fields

input_coins: Vec<ObjectID>

The input coins to be used for pay recipients, including the gas coin.

recipient: SuiAddress

The recipient address.

gas_budget: u64

Gas budget for this transaction

Pay all residual SUI coins to the recipient with input coins, after deducting the gas cost. The input coins also include the coin for gas payment, so no extra gas coin is required.

Addresses

Obtain the Addresses managed by the client.

NewAddress

Fields

key_scheme: SignatureScheme

derivation_path: Option<DerivationPath>

Generate new address and keypair with keypair scheme flag {ed25519 | secp256k1 | secp256r1} with optional derivation path, default to m/44’/784’/0’/0’/0’ for ed25519 or m/54’/784’/0’/0/0 for secp256k1 or m/74’/784’/0’/0/0 for secp256r1.

Objects

Fields

address: Option<SuiAddress>

Address owning the objects Shows all objects owned by sui client active-address if no argument is passed

Obtain all objects owned by the address

Gas

Fields

address: Option<SuiAddress>

Address owning the objects

Obtain all gas objects owned by the address.

SplitCoin

Fields

coin_id: ObjectID

Coin to Split, in 20 bytes Hex string

amounts: Option<Vec<u64>>

Specific amounts to split out from the coin

count: Option<u64>

Count of equal-size coins to split into

gas: Option<ObjectID>

ID of the gas object for gas payment, in 20 bytes Hex string If not provided, a gas object with at least gas_budget value will be selected

gas_budget: u64

Gas budget for this call

Split a coin object into multiple coins.

MergeCoin

Fields

primary_coin: ObjectID

Coin to merge into, in 20 bytes Hex string

coin_to_merge: ObjectID

Coin to be merged, in 20 bytes Hex string

gas: Option<ObjectID>

ID of the gas object for gas payment, in 20 bytes Hex string If not provided, a gas object with at least gas_budget value will be selected

gas_budget: u64

Gas budget for this call

Merge two coin objects into one coin

CreateExampleNFT

Fields

name: Option<String>

Name of the NFT

description: Option<String>

Description of the NFT

url: Option<String>

Display url(e.g., an image url) of the NFT

gas: Option<ObjectID>

ID of the gas object for gas payment, in 20 bytes Hex string If not provided, a gas object with at least gas_budget value will be selected

gas_budget: Option<u64>

Gas budget for this transfer

Create an example NFT

SerializeTransferSui

Fields

to: SuiAddress

Recipient address

sui_coin_object_id: ObjectID

Sui coin object to transfer, ID in 20 bytes Hex string. This is also the gas object.

gas_budget: u64

Gas budget for this transfer

amount: Option<u64>

The amount to transfer, if not specified, the entire coin object will be transferred.

Serialize a transfer that can be signed. This is useful when user prefers to take the data to sign elsewhere.

ExecuteSignedTx

Fields

tx_bytes: String

BCS serialized transaction data bytes without its type tag, as base-64 encoded string.

signature: String

Base64 encoded signature flag || signature || pubkey.

Execute a Signed Transaction. This is useful when the user prefers to sign elsewhere and use this command to execute.

Implementations

impl SuiClientCommands

pub async fn execute(
self,
context: &mut WalletContext
) -> Result<SuiClientCommandResult, Error>

pub fn switch_env(config: &mut SuiClientConfig, env: &str) -> Result<(), Error>