sui::client_ptb

Module builder

Structs

PTBBuilder

The PTBBuilder struct is the main workhorse that transforms a sequence of ParsedPTBCommands into an actual PTB that can be run. The main things to keep in mind are that this contains:

Functions

into_struct_tag

into_type_tag

Given a parsed type, resolve it to a TypeTag. This custom function makes usage of the custom into struct tag function that makes use of the already resolved MVR data, if needed.

is_mvr_name

Check if it’s a MVR name. MVR names must contain a / in them so as we already parsed the name, we can just check if it contains a / in it. A keystore alias cannot contain a slash, and similarly a numerical address or named address cannot have a / in it.

resolve_package

Try to resolve an ObjectID to a MovePackage

to_ordinal_contraction